Bugzilla:
https://bugzilla.redhat.com/show_bug.cgi?id=434998

Description:
The code to restart syscalls after signals depends on checking for a
negative orig_rax, and for particular negative -ERESTART* values in rax.
These fields are 64 bits and for a 32-bit task they get zero-extended,
therefore they are never negative and syscall restart behavior is lost.

Solution:
Doing sign-extension where it matters.

For orig_rax, the only time the value should be -1 but winds up as
0x0ffffffff is via a 32-bit ptrace call. So the patch changes ptrace to
sign-extend the 32-bit orig_eax value when it's stored; it doesn't
change the checks on orig_rax, though it uses the new current_syscall()
inline to better document the subtle importance of the used of
signedness there.

The rax value is stored a lot of ways and it seems hard to get them all
sign-extended at their origins. So for that, we use the
current_syscall_ret() to sign-extend it only for 32-bit tasks at the
time of the -ERESTART* comparisons.

Upstream status:
commit 40f0933d51f4cba26a5c009a26bb230f4514c1b6

Test status:
Built on all arch and tested on x86_64 using the reproducer provided in
bugzilla.

Brew build:
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=1365391

Regards,
Jerome