To: Steven Rostedt <rostedt@goodmis.org> Cc: Thomas Gleixner <tglx@linutronix.de> Subject: trace_hist.c: divide-by-zero problem (2) Parts/Attachments:    1 Shown    20 lines  Text    2 Shown    86 lines  Text ----------------------------------------
Steven,

do we really need to continuously calculate the average latency and
spend lots of time in the division function? I don't think so. It is
probably sufficient to calculate the average latency only when we
display it.

What do you think?


        Carsten.



The division function of a 64-bit divisor divided by a 64-bit dividend
to calculate the average latency may be time-consuming on a 32-bit
system. We, therefore, no longer calculate the average whenever a new
latency value is added but only when we display the histogram data.

Signed-off-by: Carsten Emde <C.Emde@osadl.org>