Subject: kvm: check need_resched() inside the irq disabled region
The missing need_resched() check inside the irq_disabled region can cause
long latencies, if a interrupt with reschedule request happens between
preempt_disable() and the local_irq_disable(). In fact the interrupts are
disabled inside prepare_guest_switch(), so the race window is rather small.

This can be further optimized, but it fixes the bug for now.

Mainline bug is fixed in a similar way.

Signed-off-by: Avi Kiviti <avi@qumranet.com>
Signed-off-by: Thomas Gleixner <tgxl@linutronix.de>