From paulmck@linux.vnet.ibm.com Thu Sep 27 00:03:19 2007 To: linux-kernel@vger.kernel.org Cc: linux-rt-users@vger.kernel.org, mingo@elte.hu, akpm@linux-foundation.org,      dipankar@in.ibm.com, josht@linux.vnet.ibm.com, tytso@us.ibm.com,      dvhltc@us.ibm.com, tglx@linutronix.de, a.p.zijlstra@chello.nl,      bunk@kernel.org, ego@in.ibm.com, oleg@tv-sign.ru, srostedt@redhat.com Subject: [PATCH RFC 3/9] RCU: Preemptible RCU
Work in progress, not for inclusion.

This patch implements a new version of RCU which allows its read-side
critical sections to be preempted. It uses a set of counter pairs
to keep track of the read-side critical sections and flips them
when all tasks exit read-side critical section. The details
of this implementation can be found in this paper -

http://www.rdrop.com/users/paulmck/RCU/OLSrtRCU.2006.08.11a.pdf

This patch was developed as a part of the -rt kernel development and
meant to provide better latencies when read-side critical sections of
RCU don't disable preemption.  As a consequence of keeping track of RCU
readers, the readers have a slight overhead (optimizations in the paper).
This implementation co-exists with the "classic" RCU implementations
and can be switched to at compiler.

Also includes RCU tracing summarized in debugfs and RCU_SOFTIRQ for
the preemptible variant of RCU.

Signed-off-by: Dipankar Sarma <dipankar@in.ibm.com>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org> (for RCU_SOFTIRQ)
Signed-off-by: Paul McKenney <paulmck@us.ibm.com>