Hi Ingo,
Apply on top of what is still in -rt.

This seems to survive a kbuild -j64 & -j512 (although with that latter the
machine goes off for a while, but does return with a kernel).

If you can spare a cycle between hacking syslets and -rt, could you have a
look at the logic this patch adds?