Subject: map read/write locks back to their readers
This patch adds a mapping from the read/write lock back to the owners
that are readers. This is a link list of tasks that own the lock for read.

The link list is protected by the read/write lock's mutex wait_lock. To
prevent grabbing this spinlock on the fast path, the list in not updated
when there is only one reader. The reader task is pointed to by the
owner field of the rw_mutex.  When the second reader grabs the read lock
it will add the first owner to the list under the wait_lock.

Signed-off-by: Steven Rostedt <srostedt@redhat.com>