From ghaskins@novell.com Fri May 23 23:34:24 2008 To: mingo@elte.hu, tglx@linutronix.de, rostedt@goodmis.org,      linux-rt-users@vger.kernel.org Cc: linux-kernel@vger.kernel.org, sdietrich@novell.com, pmorreale@novell.com,      mkohari@novell.com, ghaskins@novell.com Subject: [PATCH 4/5] adjust pi_lock usage in wakeup
    [ The following text is in the "utf-8" character set. ]
    [ Your display is set for the "iso-8859-1" character set.  ]
    [ Some characters may be displayed incorrectly. ]

From: Peter W.Morreale <pmorreale@novell.com>

In wakeup_next_waiter(), we take the pi_lock, and then find out whether
we have another waiter to add to the pending owner.  We can reduce
contention on the pi_lock for the pending owner if we first obtain the
pointer to the next waiter outside of the pi_lock.

Signed-off-by: Peter W. Morreale <pmorreale@novell.com>
Signed-off-by: Gregory Haskins <ghaskins@novell.com>